+++ umich_catalog_indexing/lib/common/subject/non_lc_subject.rb
@@ -2,11 +2,11 @@
 
 require_relative "lc_subject"
 
-module HathiTrust::Subject
+module Common::Subject
   # There are a wide variety of non-LC subject types (e.g., MESH). For the
   # moment, just treat them all the same as LC Hierarchical, with delimiters
   # between every subfield value
-  class NonLCSubject < HathiTrust::Subject::LCSubjectHierarchical
+  class NonLCSubject < Common::Subject::LCSubjectHierarchical
 
     # Here for testing purposes to distinguish from the LC subjects
     # @return [Boolean]
